requirements.txt
Django==1.8.17
markdown2==2.3.1
Pillow==4.0.0
Pygments==2.1.3
pytz==2016.10
PyMySQL==0.7.9
setuptools==20.10.1
django-tinymce==2.4.0